requirements.txt
discord.py==2.3.2
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.65
schedule==1.2.0
pytz==2024.2
pytest==8.4.1
black==25.1.0
flake8==7.3.0
isort==6.0.1